+++ lib/constants/settings.dart
@@ -1,0 +1,12 @@
+const List<String> validFileExtensions = [
+  "jpeg",
+  "jpg",
+  "png",
+  "gif",
+  "bmp",
+  "tiff",
+  "ico",
+  "webp",
+  "psd",
+  "pdf",
+];